+++ docs/changelog.md
@@ -21,7 +21,9 @@ Added:
 Fixed:
 - [TOFAnalogResponse][extra.applications.TOFAnalogResponse] has improved numerical stability (!538).
 
-Changed:
+Fixed:
+- Fixed support for future PythonCall.jl versions in
+  [imshow2()][extra.utils.imshow2] (!547).
 
 ## [2026.1.1]